+++ src/CompilerCxx/Procedure.hs
@@ -233,19 +233,18 @@ compileStatement (IgnoreValues c e) = do
   (_,e') <- compileExpression e
   maybeSetTrace c
   csWrite ["(void) (" ++ useAsWhatever e' ++ ");"]
-compileStatement (Assignment c as e) = do
+compileStatement (Assignment c as e) = flip reviseErrorStateT message $ do
   (ts,e') <- compileExpression e
   r <- csResolver
   fa <- csAllFilters
   -- Check for a count match first, to avoid the default error message.
-  _ <- processPairsT alwaysPair (fmap assignableName as) ts `reviseErrorStateT`
-    ("In assignment at " ++ formatFullContext c)
-  _ <- processPairsT (createVariable r fa) as ts `reviseErrorStateT`
-    ("In assignment at " ++ formatFullContext c)
+  _ <- processPairsT alwaysPair (fmap assignableName as) ts
+  _ <- processPairsT (createVariable r fa) as ts
   maybeSetTrace c
   variableTypes <- sequence $ map (uncurry getVariableType) $ zip (pValues as) (pValues ts)
   assignAll (zip3 ([0..] :: [Int]) variableTypes (pValues as)) e'
   where
+    message = "In assignment at " ++ formatFullContext c
     assignAll [v] e2 = assignSingle v e2
     assignAll vs e2 = do
       csWrite ["{","const auto r = " ++ useAsReturns e2 ++ ";"]
@@ -256,21 +255,21 @@ compileStatement (Assignment c as e) = do
       (VariableValue _ _ t _) <- csGetVariable c2 n
       return t
     getVariableType (ExistingVariable (DiscardInput _)) t = return t
-    createVariable r fa (CreateVariable c2 t1 n) t2 = do
-      -- TODO: Call csRequiresTypes for t1. (Maybe needs a helper function.)
-      lift $ mergeAllM [validateGeneralInstance r fa (vtType t1),
-                        checkValueTypeMatch_ r fa t2 t1] `reviseErrorM`
-        ("In creation of " ++ show n ++ " at " ++ formatFullContext c2)
-      csAddVariable c2 n (VariableValue c2 LocalScope t1 True)
-      csWrite [variableStoredType t1 ++ " " ++ variableName n ++ ";"]
-    createVariable r fa (ExistingVariable (InputValue c2 n)) t2 = do
-      (VariableValue _ _ t1 w) <- csGetVariable c2 n
-      when (not w) $ lift $ compileErrorM $ "Cannot assign to read-only variable " ++
-                                           show n ++ formatFullContextBrace c2
-      -- TODO: Also show original context.
-      lift $ (checkValueTypeMatch_ r fa t2 t1) `reviseErrorM`
-        ("In assignment to " ++ show n ++ " at " ++ formatFullContext c2)
-      csUpdateAssigned n
+    createVariable r fa (CreateVariable c2 t1 n) t2 =
+      flip reviseErrorStateT ("In creation of " ++ show n ++ " at " ++ formatFullContext c2) $ do
+        -- TODO: Call csRequiresTypes for t1. (Maybe needs a helper function.)
+        lift $ mergeAllM [validateGeneralInstance r fa (vtType t1),
+                          checkValueTypeMatch_ r fa t2 t1]
+        csAddVariable c2 n (VariableValue c2 LocalScope t1 True)
+        csWrite [variableStoredType t1 ++ " " ++ variableName n ++ ";"]
+    createVariable r fa (ExistingVariable (InputValue c2 n)) t2 =
+      flip reviseErrorStateT ("In assignment to " ++ show n ++ " at " ++ formatFullContext c2) $ do
+        (VariableValue _ _ t1 w) <- csGetVariable c2 n
+        when (not w) $ lift $ compileErrorM $ "Cannot assign to read-only variable " ++
+                                              show n ++ formatFullContextBrace c2
+        -- TODO: Also show original context.
+        lift $ (checkValueTypeMatch_ r fa t2 t1)
+        csUpdateAssigned n
     createVariable _ _ _ _ = return ()
     assignSingle (_,t,CreateVariable _ _ n) e2 =
       csWrite [variableName n ++ " = " ++ writeStoredVariable t e2 ++ ";"]
@@ -843,11 +842,10 @@ guessParamsFromArgs :: (Show c, MergeableM m, CompileErrorM m, TypeResolver r) =
   r -> ParamFilters -> ScopedFunction c -> Positional (InstanceOrInferred c) ->
   Positional ValueType -> m (Positional GeneralInstance)
 guessParamsFromArgs r fa f ps ts = do
-  let fa2 = fa `Map.union` getFunctionFilterMap f
+  let ff = getFunctionFilterMap f
   args <- processPairs alwaysPair ts (fmap pvType $ sfArgs f)
   pa <- fmap Map.fromList $ processPairs toInstance (fmap vpParam $ sfParams f) ps
-  let pa2 = pa `Map.union` (Map.fromList $ zip (Map.keys fa2) (map (SingleType . JustParamName) $ Map.keys fa2))
-  pa3 <- inferParamTypes r fa2 pa2 args
+  pa3 <- inferParamTypes r fa ff pa args
   fmap Positional $ mapErrorsM (subPosition pa3) (pValues $ sfParams f) where
     subPosition pa2 p =
       case (vpParam p) `Map.lookup` pa2 of